resolve: merge interfaces
