+++ base/src/main/java/org/aya/resolve/context/NoExportContext.java
@@ -2,7 +2,6 @@
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 package org.aya.resolve.context;
 
-import kala.collection.Map;
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableHashMap;
 import kala.collection.mutable.MutableMap;
@@ -39,8 +38,7 @@ public record NoExportContext(
   public void doExport(@NotNull ModulePath componentName, @NotNull String name, @NotNull DefVar<?, ?> ref, @NotNull SourcePos sourcePos) {
   }
 
-  @Override
-  public @NotNull Map<ModulePath, ModuleExport> exports() {
-    return Map.empty();
+  @Override public @NotNull MutableMap<ModulePath, ModuleExport> exports() {
+    return MutableMap.create();
   }
 }
